Update the test fixture to use the OSF configuration returned by getOsfServiceConfigurationResource().
In order for this to work properly, the initialization of the mock HTTP servers and their expectations had to be moved from static @BeforeClass methods to @Before methods, so that the non-static getOsfServiceConfigurationResource() method could be accessed.

In addition, in order for other test suites outside of the current classloader to specifiy a configuration, a Spring ResourceLoader abstraction was added.  This allows modules like the osf-packaging module to rely on the AbstractMockServerTest to properly resolve configuration files.