Include comments in the WikiTest unit test
The existing JSON files for WikiTest were updated to use localhost:8000
URLs and the comment section was reintroduced to the wiki JSON.  This
allowed the test to proceed where it had previously failed, and required
the inclusion of several additional files.  When the existing test
passed again, some assertions were added to evaluate the comment content
for the wiki.